requirements.txt
# chromedriver-binary==81.0.4044.69.0
selenium==3.141.0
# urllib3==1.25.8
# webdriver-manager==2.3.0
Pillow==7.1.2
requests==2.23.0
matplotlib==3.2.1